api/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
fastapi[all]>=0.119.0
langchain>=0.3.27
langchain-openai>=0.3.35
azure-cosmos>=4.6.0
python-dotenv>=1.1.1
uvicorn>=0.37.0
pydantic-settings